pyproject.toml
[project]
name = "containers"
version = "0.1.0"
description = "A repository containing various Dockerfiles."
authors = [{ name = "Marcus R. Brown", email = "<git@mrbro.dev>" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
  "argparse (>=1.4.0,<2.0.0)",
  "pyyaml (>=6.0.2,<7.0.0)",
  "requests (>=2.33.0,<3.0.0)",
  "jinja2 (>=3.0.0,<4.0.0)",
  "jsonschema (>=4.0.0,<5.0.0)",
  "openai (>=2.32.0,<2.33.0)",
  "anthropic (>=0.30.0,<1.0.0)",
]

[project.scripts]
containers = "scripts.containers_cli:main"
collect-docker-metrics = "scripts.collect_docker_metrics:main"
generate-image-tags = "scripts.generate_image_tags:main"
generate-dockerfile = "scripts.generate_dockerfile:main"
template-engine = "scripts.template_engine:main"
template-testing = "scripts.template_testing:main"
generate-docs = "scripts.template_documentation:main"
ai-chat = "scripts.ai_chat_interface:main"
ai-analyze = "scripts.template_intelligence:analyze_project"
ai-recommend = "scripts.template_intelligence:recommend_templates"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py313']
extend-exclude = '''
(
  ^/templates/
)
'''

[tool.isort]
profile = "black"
known_third_party = ["pytest"]
extend_skip = ["templates"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=scripts.ai_core --cov=scripts.template_engine --cov=scripts.containers_cli --cov=scripts.predictive_maintenance --cov-report=term-missing --cov-fail-under=35"

[tool.coverage.run]
branch = true

[tool.coverage.report]
skip_covered = false
show_missing = true

[tool.poetry]
packages = [{ include = "scripts" }]

[tool.poetry.group.dev.dependencies]
pytest = "^9.0"
pytest-cov = "^7.0"
pytest-mock = "^3.0"
black = ">=26.3.1"
isort = "^8.0"
pylint = "^4.0"
yamllint = "^1.0"
